requirements.txt
gradio>=6.12
langgraph>=0.2
langchain-groq>=0.2
langchain-core>=0.3
groq>=0.4
pydantic>=2.0
requests>=2.31
fpdf2>=2.7
huggingface_hub>=0.24
